requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-jose==3.3.0
pydantic==2.8.2
tenacity==8.3.0